+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/Governor.hs
@@ -16,6 +16,7 @@ module Ouroboros.Network.PeerSelection.Governor
   , PeerSelectionActions (..)
   , PeerStateActions (..)
   , TracePeerSelection (..)
+  , ChurnAction (..)
   , DebugPeerSelection (..)
   , DebugPeerSelectionState (..)
   , peerSelectionGovernor
@@ -29,6 +30,7 @@ module Ouroboros.Network.PeerSelection.Governor
   , PublicPeerSelectionState (..)
   , makePublicPeerSelectionStateVar
   , PeerSelectionCounters (..)
+  , emptyPeerSelectionCounters
   , nullPeerSelectionTargets
   , emptyPeerSelectionState
   , ChurnMode (..)
@@ -449,12 +451,13 @@ peerSelectionGovernor :: ( Alternative (STM m)
                       -> Tracer m (DebugPeerSelection peeraddr)
                       -> Tracer m PeerSelectionCounters
                       -> StdGen
+                      -> StrictTVar m PeerSelectionCounters
                       -> StrictTVar m (PublicPeerSelectionState peeraddr)
                       -> StrictTVar m (PeerSelectionState peeraddr peerconn)
                       -> PeerSelectionActions peeraddr peerconn m
                       -> PeerSelectionPolicy  peeraddr m
                       -> m Void
-peerSelectionGovernor tracer debugTracer countersTracer fuzzRng publicStateVar debugStateVar actions policy =
+peerSelectionGovernor tracer debugTracer countersTracer fuzzRng countersVar publicStateVar debugStateVar actions policy =
     JobPool.withJobPool $ \jobPool -> do
       localPeers <- map (\(w, h, _) -> (w, h))
                 <$> atomically (readLocalRootPeers actions)
@@ -462,6 +465,7 @@ peerSelectionGovernor tracer debugTracer countersTracer fuzzRng publicStateVar d
         tracer
         debugTracer
         countersTracer
+        countersVar
         publicStateVar
         debugStateVar
         actions
@@ -497,6 +501,7 @@ peerSelectionGovernorLoop :: forall m peeraddr peerconn.
                           => Tracer m (TracePeerSelection peeraddr)
                           -> Tracer m (DebugPeerSelection peeraddr)
                           -> Tracer m PeerSelectionCounters
+                          -> StrictTVar m PeerSelectionCounters
                           -> StrictTVar m (PublicPeerSelectionState peeraddr)
                           -> StrictTVar m (PeerSelectionState peeraddr peerconn)
                           -> PeerSelectionActions peeraddr peerconn m
@@ -507,6 +512,7 @@ peerSelectionGovernorLoop :: forall m peeraddr peerconn.
 peerSelectionGovernorLoop tracer
                           debugTracer
                           countersTracer
+                          countersVar
                           publicStateVar
                           debugStateVar
                           actions
@@ -554,10 +560,19 @@ peerSelectionGovernorLoop tracer
       let Decision { decisionTrace, decisionJobs, decisionState } =
             timedDecision now
           !newCounters = peerStateToCounters decisionState
-      traverse_ (traceWith tracer) decisionTrace
+
+      atomically $ do
+        -- Update counters
+        withCacheA (countersCache decisionState)
+                   newCounters
+                   (writeTVar countersVar)
+
+      -- Trace counters
       traceWithCache countersTracer
                      (countersCache decisionState)
                      newCounters
+
+      traverse_ (traceWith tracer) decisionTrace
 
       mapM_ (JobPool.forkJob jobPool) decisionJobs
       loop (decisionState { countersCache = Cache newCounters }) dbgUpdateAt'